# book-network-ui/src/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "OpenApi specification - Abdelrahman magdi",
        "description": "OpenApi documentation for Spring Security",
        "termsOfService": "Terms of service",
        "contact": {
            "name": "Abdelrahman magdi",
            "email": "abdelrahmanmagdi527@gmail.com"
        },
        "license": {
            "name": "Licence name",
            "url": "https://some-url.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8088/api/v1",
            "description": "Local ENV"
        },
        {
            "url": "https://aliboucoding.com/course",
            "description": "PROD ENV"
        }
    ],
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Book"
                ],
                "operationId": "findAllBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Book"
                ],
                "operationId": "saveBook",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/cover/{book-id}": {
            "post": {
                "tags": [
                    "Book"
                ],
                "operationId": "uploadBookCoverPicture",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/borrow/{book-id}": {
            "post": {
                "tags": [
                    "Book"
                ],
                "operationId": "borrowBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "request",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegistrationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/authentication": {
            "post": {
                "tags": [
                    "auth"
                ],
                "operationId": "authentication",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthenticationRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthenticationResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/shareable/{book-id}": {
            "patch": {
                "tags": [
                    "Book"
                ],
                "operationId": "updateShareableStatus",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/borrow/return/{book-id}": {
            "patch": {
                "tags": [
                    "Book"
                ],
                "operationId": "returnBorrowBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/borrow/return/approve/{book-id}": {
            "patch": {
                "tags": [
                    "Book"
                ],
                "operationId": "approveReturnBorrowBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/archived/{book-id}": {
            "patch": {
                "tags": [
                    "Book"
                ],
                "operationId": "updateArchivedStatus",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/feedbacks/book/{book-id}": {
            "get": {
                "tags": [
                    "Feedback"
                ],
                "operationId": "findAllFeedbacksByBook",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseFeedbackResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{book-id}": {
            "get": {
                "tags": [
                    "Book"
                ],
                "operationId": "findBookById",
                "parameters": [
                    {
                        "name": "book-id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/returned": {
            "get": {
                "tags": [
                    "Book"
                ],
                "operationId": "findAllReturnedBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBorrowedBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/owner": {
            "get": {
                "tags": [
                    "Book"
                ],
                "operationId": "findAllBooksByOwner",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/borrowed": {
            "get": {
                "tags": [
                    "Book"
                ],
                "operationId": "findAllBorrowedBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PageResponseBorrowedBookResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BookRequest": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string",
                        "minLength": 1
                    },
                    "authorName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "isbn": {
                        "type": "string",
                        "minLength": 1
                    },
                    "synopsis": {
                        "type": "string",
                        "minLength": 1
                    },
                    "shareable": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "authorName",
                    "isbn",
                    "synopsis",
                    "title"
                ]
            },
            "RegistrationRequest": {
                "type": "object",
                "properties": {
                    "firstname": {
                        "type": "string",
                        "minLength": 1
                    },
                    "lastname": {
                        "type": "string",
                        "minLength": 1
                    },
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 2147483647,
                        "minLength": 8
                    }
                },
                "required": [
                    "email",
                    "firstname",
                    "lastname",
                    "password"
                ]
            },
            "AuthenticationRequest": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "minLength": 1
                    },
                    "password": {
                        "type": "string",
                        "maxLength": 2147483647,
                        "minLength": 8
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "AuthenticationResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
            "FeedbackResponse": {
                "type": "object",
                "properties": {
                    "note": {
                        "type": "number",
                        "format": "double"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "ownFeedback": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseFeedbackResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FeedbackResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
            "BookResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "authorName": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "synopsis": {
                        "type": "string"
                    },
                    "owner": {
                        "type": "string"
                    },
                    "cover": {
                        "type": "string",
                        "format": "byte"
                    },
                    "rate": {
                        "type": "number",
                        "format": "double"
                    },
                    "archived": {
                        "type": "boolean"
                    },
                    "shareable": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseBookResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BookResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            },
            "BorrowedBookResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "title": {
                        "type": "string"
                    },
                    "authorName": {
                        "type": "string"
                    },
                    "isbn": {
                        "type": "string"
                    },
                    "rate": {
                        "type": "number",
                        "format": "double"
                    },
                    "returned": {
                        "type": "boolean"
                    },
                    "returnApproved": {
                        "type": "boolean"
                    }
                }
            },
            "PageResponseBorrowedBookResponse": {
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/BorrowedBookResponse"
                        }
                    },
                    "number": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalElements": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "totalPages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "first": {
                        "type": "boolean"
                    },
                    "last": {
                        "type": "boolean"
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT auth description",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}